backend/requirements.txt
# ─── Web Framework ──────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# ─── Auth ───────────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ─── Database ───────────────────────────────────────────────────────────────
motor>=3.3.0
pymongo>=4.6.0

# ─── Config ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.2.0
email-validator>=2.1.0

# ─── Multi-Format Parsing ───────────────────────────────────────────────────
pymupdf>=1.23.0        # PDF
python-docx>=1.1.0     # Word
python-pptx>=0.6.23    # PPT
pillow>=10.2.0         # Image processing
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0

# ─── AI / LLM ───────────────────────────────────────────────────────────────
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0
groq>=0.5.0

# ─── Vector Store & Embeddings ──────────────────────────────────────────────
faiss-cpu>=1.7.4
# sentence-transformers removed to save memory on Render Free Tier
numpy>=1.26.0

# ─── Utilities ──────────────────────────────────────────────────────────────
httpx>=0.27.0
aiofiles>=23.0.0
